# Improved Near Cache tests