--- a/hazelcast-client/src/main/java/com/hazelcast/client/proxy/ClientReplicatedMapProxy.java
+++ b/hazelcast-client/src/main/java/com/hazelcast/client/proxy/ClientReplicatedMapProxy.java
@@ -541,7 +541,7 @@ public class ClientReplicatedMapProxy<K, V> extends ClientProxy implements Repli
         if (nearCache == null) {
             return;
         }
-        nearCache.remove(key);
+        nearCache.invalidate(key);
     }
 
     private class ReplicatedMapEventHandler
@@ -620,7 +620,7 @@ public class ClientReplicatedMapProxy<K, V> extends ClientProxy implements Repli
                 case UPDATED:
                 case EVICTED:
                     K key = toObject(dataKey);
-                    nearCache.remove(key);
+                    nearCache.invalidate(key);
                     break;
                 case CLEAR_ALL:
                     nearCache.clear();