--- a/hazelcast/src/main/java/com/hazelcast/monitor/NearCacheStats.java
+++ b/hazelcast/src/main/java/com/hazelcast/monitor/NearCacheStats.java
@@ -83,11 +83,6 @@ public interface NearCacheStats extends LocalInstanceStats {
     long getInvalidations();
 
     /**
-     * Resets the number of invalidations of Near Cache entries owned by this member.
-     */
-    void resetInvalidations();
-
-    /**
      * Returns the number of Near Cache key persistences (when the pre-load feature is enabled).
      *
      * @return the number of Near Cache key persistences