--- a/hazelcast/src/main/java/com/hazelcast/monitor/impl/NearCacheStatsImpl.java
+++ b/hazelcast/src/main/java/com/hazelcast/monitor/impl/NearCacheStatsImpl.java
@@ -46,8 +46,8 @@ public class NearCacheStatsImpl implements NearCacheStats {
             newUpdater(NearCacheStatsImpl.class, "expirations");
     private static final AtomicLongFieldUpdater<NearCacheStatsImpl> INVALIDATIONS =
             newUpdater(NearCacheStatsImpl.class, "invalidations");
-    private static final AtomicLongFieldUpdater<NearCacheStatsImpl> RECEIVED_INVALIDATIONS =
-            newUpdater(NearCacheStatsImpl.class, "receivedInvalidations");
+    private static final AtomicLongFieldUpdater<NearCacheStatsImpl> INVALIDATION_REQUESTS =
+            newUpdater(NearCacheStatsImpl.class, "invalidationRequests");
     private static final AtomicLongFieldUpdater<NearCacheStatsImpl> PERSISTENCE_COUNT =
             newUpdater(NearCacheStatsImpl.class, "persistenceCount");
 
@@ -69,7 +69,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
     @Probe
     private volatile long invalidations;
     @Probe
-    private volatile long receivedInvalidations;
+    private volatile long invalidationRequests;
 
     @Probe
     private volatile long persistenceCount;
@@ -97,7 +97,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
         evictions = stats.evictions;
         expirations = stats.expirations;
         invalidations = stats.invalidations;
-        receivedInvalidations = stats.receivedInvalidations;
+        invalidationRequests = stats.invalidationRequests;
 
         persistenceCount = stats.persistenceCount;
         lastPersistenceTime = stats.lastPersistenceTime;
@@ -218,17 +218,16 @@ public class NearCacheStatsImpl implements NearCacheStats {
         INVALIDATIONS.addAndGet(this, delta);
     }
 
-    public long getReceivedInvalidations() {
-        return receivedInvalidations;
+    public long getInvalidationRequests() {
+        return invalidationRequests;
     }
 
-    public void incrementReceivedInvalidations() {
-        RECEIVED_INVALIDATIONS.incrementAndGet(this);
+    public void incrementInvalidationRequests() {
+        INVALIDATION_REQUESTS.incrementAndGet(this);
     }
 
-    public void resetInvalidations() {
-        INVALIDATIONS.set(this, 0);
-        RECEIVED_INVALIDATIONS.set(this, 0);
+    public void resetInvalidationEvents() {
+        INVALIDATION_REQUESTS.set(this, 0);
     }
 
     @Override
@@ -294,7 +293,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
         root.add("evictions", evictions);
         root.add("expirations", expirations);
         root.add("invalidations", invalidations);
-        root.add("receivedInvalidations", receivedInvalidations);
+        root.add("invalidationEvents", invalidationRequests);
         root.add("persistenceCount", persistenceCount);
         root.add("lastPersistenceTime", lastPersistenceTime);
         root.add("lastPersistenceDuration", lastPersistenceDuration);
@@ -314,7 +313,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
         evictions = getLong(json, "evictions", -1L);
         expirations = getLong(json, "expirations", -1L);
         invalidations = getLong(json, "invalidations", -1L);
-        receivedInvalidations = getLong(json, "receivedInvalidations", -1L);
+        invalidationRequests = getLong(json, "invalidationEvents", -1L);
         persistenceCount = getLong(json, "persistenceCount", -1L);
         lastPersistenceTime = getLong(json, "lastPersistenceTime", -1L);
         lastPersistenceDuration = getLong(json, "lastPersistenceDuration", -1L);