--- a/hazelcast/src/main/java/com/hazelcast/monitor/impl/NearCacheStatsImpl.java
+++ b/hazelcast/src/main/java/com/hazelcast/monitor/impl/NearCacheStatsImpl.java
@@ -44,6 +44,8 @@ public class NearCacheStatsImpl implements NearCacheStats {
             newUpdater(NearCacheStatsImpl.class, "evictions");
     private static final AtomicLongFieldUpdater<NearCacheStatsImpl> EXPIRATIONS =
             newUpdater(NearCacheStatsImpl.class, "expirations");
+    private static final AtomicLongFieldUpdater<NearCacheStatsImpl> INVALIDATIONS =
+            newUpdater(NearCacheStatsImpl.class, "invalidations");
     private static final AtomicLongFieldUpdater<NearCacheStatsImpl> PERSISTENCE_COUNT =
             newUpdater(NearCacheStatsImpl.class, "persistenceCount");
 
@@ -61,6 +63,9 @@ public class NearCacheStatsImpl implements NearCacheStats {
     private volatile long evictions;
     @Probe
     private volatile long expirations;
+
+    @Probe
+    private volatile long invalidations;
 
     @Probe
     private volatile long persistenceCount;
@@ -87,6 +92,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
         misses = stats.misses;
         evictions = stats.evictions;
         expirations = stats.expirations;
+        invalidations = stats.invalidations;
 
         persistenceCount = stats.persistenceCount;
         lastPersistenceTime = stats.lastPersistenceTime;
@@ -195,6 +201,24 @@ public class NearCacheStatsImpl implements NearCacheStats {
     }
 
     @Override
+    public long getInvalidations() {
+        return invalidations;
+    }
+
+    public void incrementInvalidations() {
+        INVALIDATIONS.incrementAndGet(this);
+    }
+
+    public void incrementInvalidations(long delta) {
+        INVALIDATIONS.addAndGet(this, delta);
+    }
+
+    @Override
+    public void resetInvalidations() {
+        INVALIDATIONS.set(this, 0);
+    }
+
+    @Override
     public long getPersistenceCount() {
         return persistenceCount;
     }
@@ -256,6 +280,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
         root.add("misses", misses);
         root.add("evictions", evictions);
         root.add("expirations", expirations);
+        root.add("invalidations", invalidations);
         root.add("persistenceCount", persistenceCount);
         root.add("lastPersistenceTime", lastPersistenceTime);
         root.add("lastPersistenceDuration", lastPersistenceDuration);
@@ -274,6 +299,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
         misses = getLong(json, "misses", -1L);
         evictions = getLong(json, "evictions", -1L);
         expirations = getLong(json, "expirations", -1L);
+        invalidations = getLong(json, "invalidations", -1L);
         persistenceCount = getLong(json, "persistenceCount", -1L);
         lastPersistenceTime = getLong(json, "lastPersistenceTime", -1L);
         lastPersistenceDuration = getLong(json, "lastPersistenceDuration", -1L);
@@ -293,6 +319,7 @@ public class NearCacheStatsImpl implements NearCacheStats {
                 + ", ratio=" + format("%.1f%%", getRatio())
                 + ", evictions=" + evictions
                 + ", expirations=" + expirations
+                + ", invalidations=" + invalidations
                 + ", lastPersistenceTime=" + lastPersistenceTime
                 + ", persistenceCount=" + persistenceCount
                 + ", lastPersistenceDuration=" + lastPersistenceDuration