… hazelcast-client/src/test/java/com/hazelcast/client/cache/nearcache/ClientCacheNearCacheLeakTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/cache/nearcache/ClientCacheNearCachePreloaderTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/ClientMapInvalidationListener.java
… hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/ClientMapNearCacheBasicTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/ClientMapNearCacheLeakTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/ClientMapNearCachePreloaderStressTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/ClientMapNearCachePreloaderTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/LiteMemberClientMapNearCacheBasicTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/replicatedmap/nearcache/ClientReplicatedMapInvalidationListener.java
… hazelcast-client/src/test/java/com/hazelcast/client/replicatedmap/nearcache/ClientReplicatedMapNearCacheBasicTest.java
… hazelcast-client/src/test/java/com/hazelcast/client/replicatedmap/nearcache/ClientReplicatedMapNearCacheLeakTest.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/AbstractNearCacheBasicTest.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/AbstractNearCacheLeakTest.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/AbstractNearCachePreloaderTest.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/NearCacheInvalidationListener.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/NearCacheTestContext.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/NearCacheTestContextBuilder.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/NearCacheTestSupport.java
… hazelcast/src/test/java/com/hazelcast/internal/nearcache/NearCacheTestUtils.java
… hazelcast/src/test/java/com/hazelcast/map/impl/nearcache/LiteMemberMapNearCacheBasicTest.java
… hazelcast/src/test/java/com/hazelcast/map/impl/nearcache/LiteMemberMapNearCacheLeakTest.java
… hazelcast/src/test/java/com/hazelcast/map/impl/nearcache/MapInvalidationListener.java
… hazelcast/src/test/java/com/hazelcast/map/impl/nearcache/MapNearCacheBasicTest.java
… hazelcast/src/test/java/com/hazelcast/map/impl/nearcache/MapNearCacheLeakTest.java
… hazelcast/src/test/java/com/hazelcast/map/impl/tx/TxnMapNearCacheBasicTest.java
… hazelcast/src/test/java/com/hazelcast/map/impl/tx/TxnMapNearCacheLeakTest.java
… hazelcast/src/test/java/com/hazelcast/monitor/impl/NearCacheStatsImplTest.java