The population of the data adapter in Near Cache tests was done during the creation of the `NearCacheTestContext`, because the initial population sends out invalidation messages and we had no method to properly wait for these invalidations to be processed. So we had a race between those invalidations and the following Near Cache population. This lead to a lot of test failures and the solution was to move the data structure population before the creation of the Near Cache instance. This is an unclean design, since the `createContext()` method now also has to take care about the data structure population, which should be a part of the test code.

The existing `NearCacheInvalidationListener` has the problem, that it runs concurrently to the real invalidation listener. So we cannot rely on those `received listener events` to wait (this is again racy and lead to test failures).

This PR introduces a Near Cache statistic for received Near Cache invalidations. There is a public statistic `invalidationCount` which is only increased on removed entries (so after the invalidation listener is processed -> race free). There is also an internal counter `receivedInvalidationCount` which is always increased (which is needed since the initial invalidation doesn't remove any entries). That counter is just accessible via the `NearCacheStatsImpl`, not the interface `NearCacheStats`.

The tests have been changed to use the `populateDataAdapter()` method on their own, which now waits for those initial invalidations. This resolves the race condition without mixing the responsibility of the `createContext()` method with the test code.

This should also speedup the `MapNearCacheBasicTest` a bit, since both members can be created before the data is populated, so there are no more partition migrations when the second member is started.

Some tests had to be changed, since we now don't test received invalidation events, but calls of the remove method. So we might have different numbers in some corner cases (as usual this happened in the lite member and `ReplicatedMap` setup). This might need further investigation, since we might have unwanted duplicated invalidation calls in the code.

The `ClientCacheInvalidationListener` is the only listener, which remains, since the `ClientNearCacheInvalidationTest` relies on asserting the incoming invalidation events, not the number of (removed) entries. The test is much better to read in the old version (I tried to re-write it to the `receivedInvalidationCount` stats, but that was a mess).

Fixes https://github.com/hazelcast/hazelcast/issues/12493
Fixes https://github.com/hazelcast/hazelcast/issues/11456