@Donnerbart 2018-04-26
In a previous version of this PR I disabled the invalidation batching completely, which of course gave the best test performance. But I had the impression that this put more load on the system. So the Near Cache tests were faster, but not the whole PR builder.

Before Sergej's review I switched this to decrease the invalidation batching interval to 1 second (default is 10 seconds). This makes the Near Cache tests a bit slower, but uses less system resources.

Package | Runtime (old) | Runtime (new) | Improvement
-- | -- | -- | --
com.hazelcast.map.impl.nearcache.invalidation | 1 min 15 sec | 1 min 10 sec | 6,67%
com.hazelcast.map.impl.nearcache | 10 min | 7 min 44 sec | 22,67%
com.hazelcast.internal.nearcache.impl.store | 18 ms | 31 ms | -72,22%
com.hazelcast.internal.nearcache.impl.preloader | 1.3 sec | 0.99 sec | 23,85%
com.hazelcast.internal.nearcache.impl.invalidation | 0.29 sec | 1.7 sec | -486,21%
com.hazelcast.internal.nearcache | 9.1 sec | 21 sec | -130,77%
com.hazelcast.client.replicatedmap.nearcache | 29 sec | 32 sec | -10,34%
com.hazelcast.client.map.impl.nearcache.invalidation | 9 sec | 9.6 sec | -6,67%
com.hazelcast.client.map.impl.nearcache | 5 min 37 sec | 6 min 8 sec | -9,20%
com.hazelcast.client.cache.nearcache | 1 min 44 sec | 2 min 37 sec | -50,96%
com.hazelcast.client.cache.impl.nearcache.invalidation | 16 sec | 38 sec | -137,50%
com.hazelcast.client.cache.impl.nearcache | 1 min 5 sec | 1 min 20 sec | -23,08%

The total improvement is 3 seconds (0,27%), so basically we didn't lose any time by waiting for the initial Near Cache invalidations (instead of starting the second instance after the initial population). I think it's still a good trade for the cleaner design.

The configuration is pulled out into the `NearCacheTestUtils`, so we can easily play around with this in the future:
```java
public static Config getBaseConfig() {
  return smallInstanceConfig()  
    .setProperty(CACHE_INVALIDATION_MESSAGE_BATCH_FREQUENCY_SECONDS.getName(), "1")
    .setProperty(MAP_INVALIDATION_MESSAGE_BATCH_FREQUENCY_SECONDS.getName(), "1");
}
```