@Donnerbart 2018-04-26
This needs an EE PR, please don't merge until this is done!