@Holmistr 2018-04-27
@Donnerbart There are some test failures on PR builder that I haven't seen before, see: https://hazelcast-l337.ci.cloudbees.com/job/new-lab-fast-pr/15394/
Let me re-run to see how often they appear.