@Holmistr 2018-05-02
@ahmetmircik @taburet @Donnerbart So are we good to merge this?