@Donnerbart 2018-05-09
@ahmetmircik Are you okay with the PR after my last commit?